+++ src/main/java/net/frozenblock/wilderwild/entity/variant/moobloom/MoobloomVariants.java
@@ -17,23 +17,27 @@
 
 package net.frozenblock.wilderwild.entity.variant.moobloom;
 
+import java.util.ArrayList;
 import java.util.List;
 import net.frozenblock.wilderwild.WWConstants;
 import net.frozenblock.wilderwild.registry.WWBlocks;
 import net.frozenblock.wilderwild.registry.WilderWildRegistries;
-import net.frozenblock.wilderwild.tag.WWBiomeTags;
+import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
 import net.minecraft.core.RegistryAccess;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.TagKey;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.RandomPatchConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.SimpleBlockConfiguration;
 import org.jetbrains.annotations.NotNull;
 import net.minecraft.Util;
 
@@ -63,48 +67,71 @@ public final class MoobloomVariants {
 		@NotNull BootstrapContext<MoobloomVariant> bootstrapContext,
 		ResourceKey<MoobloomVariant> resourceKey,
 		@NotNull Block flowerBlock,
-		String textureName,
-		TagKey<Biome> biomes
+		String textureName
 	) {
 		ResourceLocation textureLocation = WWConstants.id("entity/moobloom/" + textureName);
 		bootstrapContext.register(
 			resourceKey,
 			new MoobloomVariant(
 				flowerBlock.defaultBlockState(),
-				textureLocation,
-				bootstrapContext.lookup(Registries.BIOME).getOrThrow(biomes)
+				textureLocation
 			)
 		);
 	}
 
-	public static Holder<MoobloomVariant> getSpawnVariant(@NotNull RegistryAccess registryAccess, Holder<Biome> holder, RandomSource random) {
+	public static Holder<MoobloomVariant> getSpawnVariant(@NotNull RegistryAccess registryAccess, Holder<Biome> holder, BlockPos pos, RandomSource random) {
 		Registry<MoobloomVariant> registry = registryAccess.registryOrThrow(WilderWildRegistries.MOOBLOOM_VARIANT);
+
+		List<ConfiguredFeature<?, ?>> flowerFeatures = holder.value().getGenerationSettings().getFlowerFeatures();
+		List<BlockState> biomeFlowerStates = new ArrayList<>();
+		for (ConfiguredFeature<?, ?> feature : flowerFeatures) {
+			if (feature.config() instanceof RandomPatchConfiguration randomPatchConfiguration) {
+				ConfiguredFeature<?, ?> configuredPlacedFeature = randomPatchConfiguration.feature().value().feature().value();
+				biomeFlowerStates.addAll(getBlockStatesFromConfiguredFeature(configuredPlacedFeature, pos, random));
+			}
+		}
+
 		List<Holder.Reference<MoobloomVariant>> variants = registry.holders()
-			.filter(reference -> (reference.value()).biomes().contains(holder))
+			.filter(reference -> biomeFlowerStates.contains(reference.value().getFlowerBlockState()))
 			.toList();
 
-		if (!variants.isEmpty()) {
-			return Util.getRandom(variants, random);
-		} else {
-			return registry.getRandom(random).orElseThrow();
+		if (!variants.isEmpty()) return Util.getRandom(variants, random);
+		return registry.getRandom(random).orElseThrow();
+	}
+
+	private static @NotNull List<BlockState> getBlockStatesFromConfiguredFeature(@NotNull ConfiguredFeature<?, ?> feature, BlockPos pos, RandomSource random) {
+		List<BlockState> blockStates = new ArrayList<>();
+
+		FeatureConfiguration config = feature.config();
+		if (config instanceof RandomPatchConfiguration randomPatchConfiguration) {
+			ConfiguredFeature<?, ?> configuredPlacedFeature = randomPatchConfiguration.feature().value().feature().value();
+			blockStates.addAll(getBlockStatesFromConfiguredFeature(configuredPlacedFeature, pos, random));
+		} else if (config instanceof SimpleBlockConfiguration simpleBlockConfiguration) {
+			blockStates.add(simpleBlockConfiguration.toPlace().getState(random, pos));
 		}
+
+		for (ConfiguredFeature<?, ?> nestedFeature : config.getFeatures().toList()) {
+			blockStates.addAll(getBlockStatesFromConfiguredFeature(nestedFeature, pos, random));
+		}
+
+		return blockStates;
 	}
 
 	public static void bootstrap(BootstrapContext<MoobloomVariant> bootstrapContext) {
-		register(bootstrapContext, DANDELION, Blocks.DANDELION, "moobloom_dandelion", WWBiomeTags.MOOBLOOM_DANDELION);
-		register(bootstrapContext, POPPY, Blocks.POPPY, "moobloom_poppy", WWBiomeTags.MOOBLOOM_POPPY);
-		register(bootstrapContext, AZURE_BLUET, Blocks.AZURE_BLUET, "moobloom_azure_bluet", WWBiomeTags.MOOBLOOM_AZURE_BLUET);
-		register(bootstrapContext, RED_TULIP, Blocks.RED_TULIP, "moobloom_red_tulip", WWBiomeTags.MOOBLOOM_RED_TULIP);
-		register(bootstrapContext, ORANGE_TULIP, Blocks.ORANGE_TULIP, "moobloom_orange_tulip", WWBiomeTags.MOOBLOOM_ORANGE_TULIP);
-		register(bootstrapContext, WHITE_TULIP, Blocks.WHITE_TULIP, "moobloom_white_tulip", WWBiomeTags.MOOBLOOM_WHITE_TULIP);
-		register(bootstrapContext, PINK_TULIP, Blocks.PINK_TULIP, "moobloom_pink_tulip", WWBiomeTags.MOOBLOOM_PINK_TULIP);
-		register(bootstrapContext, OXEYE_DAISY, Blocks.OXEYE_DAISY, "moobloom_oxeye_daisy", WWBiomeTags.MOOBLOOM_OXEYE_DAISY);
-		register(bootstrapContext, CORNFLOWER, Blocks.CORNFLOWER, "moobloom_cornflower", WWBiomeTags.MOOBLOOM_CORNFLOWER);
-		register(bootstrapContext, LILY_OF_THE_VALLEY, Blocks.LILY_OF_THE_VALLEY, "moobloom_lily_of_the_valley", WWBiomeTags.MOOBLOOM_LILY_OF_THE_VALLEY);
-		register(bootstrapContext, ALLIUM, Blocks.ALLIUM, "moobloom_allium", WWBiomeTags.MOOBLOOM_ALLIUM);
-		register(bootstrapContext, BLUE_ORCHID, Blocks.BLUE_ORCHID, "moobloom_blue_orchid", WWBiomeTags.MOOBLOOM_BLUE_ORCHID);
-		register(bootstrapContext, CARNATION, WWBlocks.CARNATION, "moobloom_carnation", WWBiomeTags.MOOBLOOM_CARNATION);
-		register(bootstrapContext, MARIGOLD, WWBlocks.MARIGOLD, "moobloom_marigold", WWBiomeTags.MOOBLOOM_MARIGOLD);
-		register(bootstrapContext, PASQUEFLOWER, WWBlocks.PASQUEFLOWER, "moobloom_pasqueflower", WWBiomeTags.MOOBLOOM_PASQUEFLOWER);
+		register(bootstrapContext, DANDELION, Blocks.DANDELION, "moobloom_dandelion");
+		register(bootstrapContext, POPPY, Blocks.POPPY, "moobloom_poppy");
+		register(bootstrapContext, AZURE_BLUET, Blocks.AZURE_BLUET, "moobloom_azure_bluet");
+		register(bootstrapContext, RED_TULIP, Blocks.RED_TULIP, "moobloom_red_tulip");
+		register(bootstrapContext, ORANGE_TULIP, Blocks.ORANGE_TULIP, "moobloom_orange_tulip");
+		register(bootstrapContext, WHITE_TULIP, Blocks.WHITE_TULIP, "moobloom_white_tulip");
+		register(bootstrapContext, PINK_TULIP, Blocks.PINK_TULIP, "moobloom_pink_tulip");
+		register(bootstrapContext, OXEYE_DAISY, Blocks.OXEYE_DAISY, "moobloom_oxeye_daisy");
+		register(bootstrapContext, CORNFLOWER, Blocks.CORNFLOWER, "moobloom_cornflower");
+		register(bootstrapContext, LILY_OF_THE_VALLEY, Blocks.LILY_OF_THE_VALLEY, "moobloom_lily_of_the_valley");
+		register(bootstrapContext, ALLIUM, Blocks.ALLIUM, "moobloom_allium");
+		register(bootstrapContext, BLUE_ORCHID, Blocks.BLUE_ORCHID, "moobloom_blue_orchid");
+		register(bootstrapContext, CARNATION, WWBlocks.CARNATION, "moobloom_carnation");
+		register(bootstrapContext, MARIGOLD, WWBlocks.MARIGOLD, "moobloom_marigold");
+		register(bootstrapContext, PASQUEFLOWER, WWBlocks.PASQUEFLOWER, "moobloom_pasqueflower");
 	}
 }