+++ CHANGELOG.md
@@ -20,8 +20,8 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
   - The `wilderwild:feature/has_flower_forest_flowers` biome tag has been added, differentiating Flower Field and Flower Forest flower generation.
 - Updated flower generation in Flower Fields and Flower Forests, being more visually pleasing.
 - Toned down flower generation in Tundras.
-- Blue Orchid Mooblooms no longer spawn in Flower Forests.
-- Marigold Mooblooms now spawn in Flower Fields and Flower Forests.
+- Moobloom variants are now selected based upon the flower features in the biome.
+  - As a result of this, all biome tags relating to Moobloom variants have been removed.
 - Added Gabbro blocks to Sculk-related tags they were missing from.
 - Nether Wart, Nether Wart Blocks, and Warped Wart Blocks now kill Termites upon contact.
 - Added a few missing Crimson and Warped blocks to the `wilderwild:kills_termite` block tag.